+++ server/src/main/java/com/sealll/application/reply/controller/ReplyController.java
@@ -4,6 +4,9 @@ import com.github.pagehelper.PageInfo;
 import com.sealll.application.reply.bean.Reply;
 import com.sealll.application.reply.service.ReplyService;
 import com.sealll.bean.Msg;
+import com.sealll.constant.ParameterConstants;
+import com.sealll.utils.SelfChecker;
+import org.apache.shiro.SecurityUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -22,6 +25,15 @@ public class ReplyController {
 
     @PostMapping
     public Msg addReply(@RequestBody Reply reply) {
+        String uid = reply.getUser().getUid();
+        Object o = SecurityUtils.getSubject().getSession().getAttribute(ParameterConstants.UID_SESSION_KEY);
+        if(o == null){
+            return Msg.auth("先登录");
+        }else{
+            if(!((String)o).equals(uid)){
+                return Msg.permit("没有权限");
+            }
+        }
         boolean b = replyService.addReply(reply);
         if (b) {
             return Msg.success("回复成功");
@@ -33,6 +45,10 @@ public class ReplyController {
     //    public boolean deleteReply(String reid);
     @DeleteMapping
     public Msg deleteReply(@RequestParam("reid") String reid,@RequestParam("uid")String uid,@RequestParam("pid")String pid) {
+        boolean check = SelfChecker.check(uid);
+        if(!check){
+            return Msg.permit("无权限");
+        }
         boolean b = replyService.delete2Reply(reid,uid,pid);
         if (b) {
             return Msg.success("删除成功");
